backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
pip==25.0.1
requests==2.32.3
setuptools==68.2.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
wheel==0.41.2
django-cors-headers

